Support for asm instructions alignment using NOP codes.
It adds a possibility to specify how many NOP codes to add after injection. It is useful when the injecting code is of different size than the original instruction. In this case it needs to replace the tail of the original instruction with NOP codes.

This is an original line for DLLInject:
CALL func_name 0052E264

This is same line with 3 NOP codes added after injection:
CALL func_name 0052E264 3